update sensorcontrolled spin to use compass
